fix(blevm-prover): add rsp helper
